+++ crewing/urls.py
@@ -8,8 +8,8 @@ urlpatterns = [
     path('seaman/card/<int:seaman_id>/', views.seamancard, name='seamancard'),
     path('seaman/add/', views.seaman_add, name='seaman_add'),
     path('seaman/edit/<int:seaman_id>/', views.seaman_edit, name='seaman_edit'),
-    path('seaman/<int:seaman_id>/opinion/add/', views.opinion_add, name='opinion_add'),
-    path('seaman/<int:seaman_id>/opinion/edit/<int:opinion_id>/', views.opinion_edit, name='opinion_edit'),
+    # path('seaman/<int:seaman_id>/opinion/add/', views.opinion_add, name='opinion_add'),
+    # path('seaman/<int:seaman_id>/opinion/edit/<int:opinion_id>/', views.opinion_edit, name='opinion_edit'),
     path('seaman/<int:seaman_id>/contract/add/', views.contract_add, name='contract_add'),
     path('seaman/<int:seaman_id>/contract/edit/<int:contract_id>/', views.contract_edit, name='contract_edit'),
     path('rank/list/', views.rank_list, name='rank_list'),
