+++ opinion/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Opinion
+
+
+admin.site.register(Opinion)
