+++ templates/crewing/seamancard.html
@@ -26,7 +26,7 @@
         {% include 'crewing/_seamancard_main.html' %}
       </div>
       <div role="tabpanel" class="tab-pane" id="opinion">
-        {% include 'crewing/_seamancard_opinion.html' %}
+        {% include 'opinion/_seamancard_opinion.html' %}
       </div>
       <div role="tabpanel" class="tab-pane" id="seaman-360">
         {% include 'crewing/_seamancard_360.html' %}
